Unify all cart validation in one function
