+++ srv/wordpress/wp-content/plugins/acore-wp-plugin/src/hooks/woocommerce/woocommerce.php
@@ -10,6 +10,7 @@ require_once "item_send.php";
 require_once "char_change.php";
 require_once "char_transfer.php";
 require_once "carbon_copy.php";
+require_once "cart_validation.php";
 
 // Add WooCommerce customer username to edit/view order admin page
 add_action('woocommerce_admin_order_data_after_billing_address', __NAMESPACE__ . '\woo_display_order_username', 10, 1);
